ACMS-000: Use pre-built Cohesion artifact on CI
