+++ .gitattributes
@@ -5,5 +5,6 @@
 /grumphp.yml export-ignore
 /hooks export-ignore
 /phpcs.xml.dist export-ignore
+/phpunit.xml export-ignore
 /tests export-ignore
 /.travis.yml export-ignore
