+++ composer.json
@@ -47,7 +47,8 @@
         "drupal/core-dev": "^8.9",
         "drush/drush": "^10",
         "oomphinc/composer-installers-extender": "^1.1",
-        "vijaycs85/drupal-quality-checker": "^1.2"
+        "vijaycs85/drupal-quality-checker": "^1.2",
+        "weitzman/drupal-test-traits": "^1.5"
     },
     "config": {
         "preferred-install": {
@@ -139,7 +140,16 @@
             "ln -s -f $PWD/modules $MODULES_DIR",
             "ln -s -f $PWD/src $PROFILE_DIR",
             "ln -s -f $PWD/tests $PROFILE_DIR",
-            "find ./docroot/modules -type d -name tests -prune -exec rm -r -f '{}' ';'"
+            "find ./docroot/modules -type d -name tests -prune -exec rm -r -f '{}' ';'",
+            "cp -f phpunit.xml ./docroot/core"
+        ],
+        "cohesion:fixture": [
+            "@putenv FILES_DIR=./docroot/sites/default/files",
+            "drush config:export --yes --quiet --destination $TMPDIR/cex",
+            "mkdir -p $FILES_DIR/cohesion/config",
+            "cp -f $TMPDIR/cex/cohesion_* $FILES_DIR/cohesion/config",
+            "tar -c -z -f ./tests/cohesion.tar.gz --directory $FILES_DIR cohesion",
+            "rm -rf $TMPDIR/cex"
         ],
         "nuke": "rm -r -f docroot vendor"
     }
