+++ tests/src/ExistingSite/InstallStateTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Drupal\Tests\acquia_cms\ExistingSite;
+
+use Drupal\Core\Config\ImmutableConfig;
+use weitzman\DrupalTestTraits\ExistingSiteBase;
+
+/**
+ * Tests config values that are set at install.
+ *
+ * @group acquia_cms
+ */
+class InstallStateTest extends ExistingSiteBase {
+
+  /**
+   * Assert that all install tasks have done what they should do.
+   *
+   * See acquia_cms_install_tasks().
+   */
+  public function testConfig() {
+    // Check that the default and admin themes are set as expected.
+    $theme_config = $this->config('system.theme');
+    $this->assertSame('cohesion_theme', $theme_config->get('default'));
+    $this->assertSame('claro', $theme_config->get('admin'));
+
+    // Check that the node create form is using the admin theme.
+    $this->assertTrue($this->config('node.settings')->get('use_admin_theme'));
+  }
+
+  /**
+   * Returns a config object by name.
+   *
+   * @param string $name
+   *   The name of the config object to return.
+   *
+   * @return \Drupal\Core\Config\ImmutableConfig
+   *   The config object, read-only to discourage this test from making any
+   *   changes.
+   */
+  private function config(string $name) : ImmutableConfig {
+    return $this->container->get('config.factory')->get($name);
+  }
+
+}
