+++ tests/travis/install.sh
@@ -20,6 +20,17 @@ source ../../../orca/bin/travis/_includes.sh
 # If there is no fixture, there's nothing else for us to do.
 [[ -d "$ORCA_FIXTURE_DIR" ]] || exit 0
 
+# Install dev dependencies.
+composer -d $ORCA_FIXTURE_DIR require --dev weitzman/drupal-test-traits
+
+# If there is a pre-built archive of code, assets, and templates for
+# Cohesion, import that instead of calling out to Cohesion's API.
+if [ ! -z $COHESION_ARTIFACT ] && [ -f $COHESION_ARTIFACT ]; then
+  cd "$ORCA_FIXTURE_DIR/docroot/sites/default"
+  tar -x -z -f $COHESION_ARTIFACT --directory files
+  drush cim --yes --quiet --partial --source sites/default/files/cohesion/config
+fi
+
 # In order for PHPUnit tests belonging to profile modules to even be
 # runnable, the profile's modules need to be symlinked into the
 # sites/all/modules directory. This is a long-standing limitation of
